requirements.txt
langgraph>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.1
pydantic>=2.7.0
